Move Pysa server to a separate file
